requirements.txt
fastapi == 0.115.5
uvicorn == 0.32.1
google-cloud-firestore == 2.19.0
SMTPEmail == 0.4.2
python-multipart == 0.0.19
pydantic[email]
